Refactor order repository, update to latest Core changes
ISSUE: PLCORE20-17
