+++ PacklinkPro/Ui/Component/Listing/Column/OrderDraft.php
@@ -14,10 +14,9 @@ use Magento\Ui\Component\Listing\Columns\Column;
 use Packlink\PacklinkPro\Bootstrap;
 use Packlink\PacklinkPro\Helper\UrlHelper;
 use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\Configuration;
-use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\Order\Interfaces\OrderRepository;
+use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\OrderShipmentDetails\OrderShipmentDetailsService;
 use Packlink\PacklinkPro\IntegrationCore\Infrastructure\ServiceRegister;
 use Packlink\PacklinkPro\Services\BusinessLogic\ConfigurationService;
-use Packlink\PacklinkPro\Services\BusinessLogic\OrderRepositoryService;
 
 /**
  * Class OrderDraft
@@ -70,10 +69,6 @@ class OrderDraft extends Column
      * @param array $dataSource
      *
      * @return array
-     *
-     * @throws \Magento\Framework\Exception\LocalizedException
-     * @throws \Packlink\PacklinkPro\IntegrationCore\Infrastructure\ORM\Exceptions\QueryFilterInvalidParamException
-     * @throws \Packlink\PacklinkPro\IntegrationCore\Infrastructure\ORM\Exceptions\RepositoryNotRegisteredException
      */
     public function prepareDataSource(array $dataSource)
     {
@@ -85,17 +80,17 @@ class OrderDraft extends Column
                 return $dataSource;
             }
 
-            /** @var OrderRepositoryService $orderRepositoryService */
-            $orderRepositoryService = ServiceRegister::getService(OrderRepository::CLASS_NAME);
+            /** @var OrderShipmentDetailsService $orderShipmentDetailsService */
+            $orderShipmentDetailsService = ServiceRegister::getService(OrderShipmentDetailsService::CLASS_NAME);
 
             $fieldName = $this->getData('name');
             foreach ($dataSource['data']['items'] as &$item) {
-                $orderDetails = $orderRepositoryService->getOrderDetailsById((int)$item['entity_id']);
-                if ($orderDetails === null || $orderDetails->getShipmentReference() === null) {
+                $orderDetails = $orderShipmentDetailsService->getDetailsByOrderId($item['entity_id']);
+                if ($orderDetails === null || $orderDetails->getReference() === null) {
                     continue;
                 }
 
-                $reference = $orderDetails->getShipmentReference();
+                $reference = $orderDetails->getReference();
 
                 $logoUrl = $this->assetRepo->getUrl('Packlink_PacklinkPro::images/logo.png');
