+++ PacklinkPro/etc/module.xml
@@ -8,5 +8,5 @@
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Packlink_PacklinkPro" setup_version="1.0.1" schema_version="1.0.1"/>
+    <module name="Packlink_PacklinkPro" setup_version="1.1.0" schema_version="1.1.0"/>
 </config>
